Changes to be committed: 	modified:   .vscode/settings.json 	modified:   Humspot/src/App.tsx 	new file:   Humspot/src/components/Activity/ActivityAddCommentBox.tsx 	new file:   Humspot/src/components/Activity/ActivityCommentsSection.tsx 	new file:   Humspot/src/components/Activity/ActivityDateTimeLocation.tsx 	new file:   Humspot/src/components/Activity/ActivityFavoriteButton.tsx 	new file:   Humspot/src/components/Activity/ActivityLocationMap.tsx 	new file:   Humspot/src/components/Activity/ActivityVisitedButton.tsx 	modified:   Humspot/src/elements/CarouselEntry.tsx 	modified:   Humspot/src/elements/CarouselEntrySecondary.tsx 	renamed:    Humspot/src/elements/MainCarousel.tsx -> Humspot/src/elements/template.tsx 	renamed:    Humspot/src/pages/attraction.css -> Humspot/src/pages/ActivityPage.css 	new file:   Humspot/src/pages/ActivityPage.tsx 	deleted:    Humspot/src/pages/attraction.tsx
