+++ .vscode/settings.json
@@ -1,5 +1,13 @@
 {
   "cSpell.words": [
     "humspot"
-  ]
+  ],
+  "[typescript]": {
+    "editor.defaultFormatter": "vscode.typescript-language-features"
+},
+"[typescriptreact]": {
+    "editor.defaultFormatter": "esbenp.prettier-vscode",
+    "editor.formatOnSave": true
+}, 
+
 }
