+++ Humspot/src/pages/ActivityPage.css
@@ -1,5 +1,4 @@
-.locationmap {
-}
+
 
 .locationcard {
   display: flex;
@@ -24,16 +23,12 @@
 
 .commentusername {
   font-size: medium;
-}
-
-.commentbox {
-  /* display: flex; */
-  /* align-items: start; */
+  margin-top: 5px;
 }
 
 .commentlist {
   overflow: scroll;
-  height: 40vh;
+  max-height: 35vh;
 }
 
 .commentavatar {
@@ -41,6 +36,8 @@
   height: 50px; /* Adjust the height to your desired size */
   flex-shrink: 0; /* Prevents the avatar from shrinking */
   margin-right: 10px; /* Adds spacing between the avatar and content */
+  margin-left: 5px;
+  margin-top: 5px;
 }
 
 .commentdate {
